+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/optimization/primitive/OnlineLinearInitialization.java
@@ -45,8 +45,8 @@ public class OnlineLinearInitialization implements OnlineOptimization {
                 solution.allocate(solution.demandsFree().orderedLinesStream().findFirst().orElseThrow()
                         , solution.suppliesFree().orderedLinesStream().findFirst().orElseThrow());
             } else {
-                solution.allocate(solution.demandsFree().line(0)
-                        , solution.suppliesFree().line(0));
+                solution.allocate(solution.demandsFree().chooseLine(0)
+                        , solution.suppliesFree().chooseLine(0));
             }
         }
     }