+++ projects/net.splitcells.gel.sheath/src/main/java/net/splitcells/gel/rating/rater/MinimalDistanceTest.java
@@ -58,23 +58,23 @@ public class MinimalDistanceTest {
         testSubject.optimize
                 (optimizationEvent
                         (StepType.REMOVAL
-                                , testSubject.demands().line(1).toLinePointer()
-                                , testSubject.supplies().line(1).toLinePointer()));
+                                , testSubject.demands().chooseLine(1).toLinePointer()
+                                , testSubject.supplies().chooseLine(1).toLinePointer()));
         testSubject.optimize
                 (optimizationEvent
                         (StepType.REMOVAL
-                                , testSubject.demands().line(2).toLinePointer()
-                                , testSubject.supplies().line(2).toLinePointer()));
+                                , testSubject.demands().chooseLine(2).toLinePointer()
+                                , testSubject.supplies().chooseLine(2).toLinePointer()));
         testSubject.optimize
                 (optimizationEvent
                         (StepType.ADDITION
-                                , testSubject.demands().line(1).toLinePointer()
-                                , testSubject.supplies().line(2).toLinePointer()));
+                                , testSubject.demands().chooseLine(1).toLinePointer()
+                                , testSubject.supplies().chooseLine(2).toLinePointer()));
         testSubject.optimize
                 (optimizationEvent
                         (StepType.REMOVAL
-                                , testSubject.demands().line(1).toLinePointer()
-                                , testSubject.supplies().line(2).toLinePointer()));
+                                , testSubject.demands().chooseLine(1).toLinePointer()
+                                , testSubject.supplies().chooseLine(2).toLinePointer()));
 
     }